fix: refactoring JiraOAuth
